my_pwd & my_echo fixed
