+++ .vscode/configurationCache.log
@@ -1,0 +1,1 @@
+{"buildTargets":["all","clean","fclean","libft/libft.a"],"launchTargets":["/Users/ilahyani/Minishell>minishell()"],"customConfigurationProvider":{"workspaceBrowse":{"browsePath":["/Users/ilahyani/Minishell"],"compilerArgs":["-Wall","-Wextra","-Werror","-lreadline","./libft/libft.a","builtins.c","main.c","utils.c","-o","minishell"],"compilerPath":"/usr/bin/cc","windowsSdkVersion":""},"fileIndex":[["/Users/ilahyani/Minishell/*.c",{"uri":{"$mid":1,"fsPath":"/Users/ilahyani/Minishell/*.c","path":"/Users/ilahyani/Minishell/*.c","scheme":"file"},"configuration":{"defines":[],"includePath":[],"forcedInclude":[],"compilerPath":"/usr/bin/cc","compilerArgs":["-Wall","-Wextra","-Werror","-lreadline","./libft/libft.a","builtins.c","main.c","utils.c","-o","minishell"],"windowsSdkVersion":""},"compileCommand":{"command":"cc -Wall -Wextra -Werror -lreadline ./libft/libft.a *.c -o minishell","directory":"/Users/ilahyani/Minishell","file":"/Users/ilahyani/Minishell/*.c"}}]]}}
