+++ my_exit.c
@@ -1,0 +1,26 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   my_exit.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: ilahyani <ilahyani@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/06/26 12:18:35 by ilahyani          #+#    #+#             */
+/*   Updated: 2022/06/26 12:18:49 by ilahyani         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+void    my_exit(char **data)
+{
+    if (sizeof_array(data)  == 1)
+        exit(1);
+    else
+    {
+        if (is_int(data[1]) && data[2])
+            printf("too many arguments\n");
+        else
+            exit(1);
+    }
+}
